feat: 로그인 로그아웃 버튼 추가, Intro 화면과 로그인 화면 분리
