+++ beer_community/src/main/frontend/src/components/Nav/Nav.jsx
@@ -2,7 +2,12 @@ import React from "react";
 import styles from "./nav.module.css";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { faAngleLeft } from "@fortawesome/free-solid-svg-icons";
+import { faArrowRightToBracket } from "@fortawesome/free-solid-svg-icons";
+import { faArrowRightFromBracket } from "@fortawesome/free-solid-svg-icons";
+import { useState } from "react";
+
 const Nav = (props) => {
+  const [loginState, setLoginState] = useState(true);
   return (
     <>
       <div className={styles.main}>
@@ -14,6 +19,14 @@ const Nav = (props) => {
               props.navigate(-1);
             }}
           />
+          <FontAwesomeIcon
+            icon={loginState===true?faArrowRightToBracket:faArrowRightFromBracket}
+            size="1x"
+            onClick={() => {
+              props.navigate("/login");
+            }}
+          />
+
         </div>
       </div>
     </>
