+++ beer_community/src/main/frontend/src/components/Nav/nav.module.css
@@ -9,6 +9,9 @@
 }
 .back {
   margin-left: 20px;
+  margin-right: 20px;
   margin-top: 20px;
   cursor: pointer;
+  display: flex;
+  justify-content: space-between;
 }
